+++ geolocator_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.2.0
+
+- Includes `satelliteCount` and `satellitesUsedInFix` in `Position`.
+
 ## 4.1.1
 
 - Updates dependencies to latest versions to prevent conflicts with other packages.